Show icon of package associated with Toast
For all those times you have some random app or background service
that posts a Toast and you have no idea who's posting it.  This adds
an icon badge to the top left corner of the Toast to show the app's
icon the Toast belongs to.

Change-Id: I82bf23664eea134f3b1f89ad5a99f6be73906ba8

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Update toast icon to properly work with 8.1

Change-Id: Ia010e7e2670fc22cf81100ad4e18a4b737200092

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Toast: Remove horrible shadow in text view

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>